+++ src/qt/walletframe.h
@@ -8,12 +8,17 @@
 #define WALLETFRAME_H
 
 #include <QFrame>
+#include <QMap>
 
 class BitcoinGUI;
 class ClientModel;
 class SendCoinsRecipient;
 class WalletModel;
-class WalletStack;
+class WalletView;
+
+QT_BEGIN_NAMESPACE
+class QStackedWidget;
+QT_END_NAMESPACE
 
 class WalletFrame : public QFrame
 {
@@ -35,7 +40,12 @@ public:
     void showOutOfSyncWarning(bool fShow);
 
 private:
-    WalletStack *walletStack;
+    QStackedWidget *walletStack;
+    BitcoinGUI *gui;
+    ClientModel *clientModel;
+    QMap<QString, WalletView*> mapWalletViews;
+
+    bool bOutOfSync;
 
 public slots:
     /** Switch to overview (home) page */
